+++ Client/Tests/Restaurant.Core.UnitTests/AutoDomainDataAttribute.cs
@@ -1,0 +1,26 @@
+﻿using System.Linq;
+using AutoFixture;
+using AutoFixture.NUnit3;
+
+namespace Restaurant.Core.UnitTests
+{
+    public class AutoDomainDataAttribute : AutoDataAttribute
+    {
+        public AutoDomainDataAttribute() : base(CustomFixture)
+        {
+        }
+
+        private static IFixture CustomFixture()
+        {
+            var fixture = new Fixture();
+
+            fixture.Behaviors.OfType<ThrowingRecursionBehavior>()
+                .ToList()
+                .ForEach(b => fixture.Behaviors.Remove(b));
+
+            fixture.Behaviors.Add(new OmitOnRecursionBehavior());
+
+            return fixture;
+        }
+    }
+}
