+++ Client/Tests/Restaurant.Core.UnitTests/ViewModels/FoodDetailViewModelTest.cs
@@ -6,7 +6,7 @@ using Restaurant.Abstractions.ViewModels;
 using Restaurant.Common.DataTransferObjects;
 using Restaurant.Core.ViewModels;
 
-namespace Restaurant.Client.UnitTests.ViewModels
+namespace Restaurant.Core.UnitTests.ViewModels
 {
 	[TestFixture]
 	public class FoodDetailViewModelTest : BaseAutoMockedTest<FoodDetailViewModel>
