+++ Client/Tests/Restaurant.Core.UnitTests/ViewModels/FoodsViewModelTests.cs
@@ -10,7 +10,7 @@ using Restaurant.Abstractions.ViewModels;
 using Restaurant.Common.DataTransferObjects;
 using Restaurant.Core.ViewModels;
 
-namespace Restaurant.Client.UnitTests.ViewModels
+namespace Restaurant.Core.UnitTests.ViewModels
 {
 	[TestFixture]
 	public class FoodsViewModelTests : BaseAutoMockedTest<FoodsViewModel>
