+++ Client/Tests/Restaurant.Core.UnitTests/ViewModels/SignInViewModelTest.cs
@@ -9,7 +9,7 @@ using Restaurant.Abstractions.ViewModels;
 using Restaurant.Common.DataTransferObjects;
 using Restaurant.Core.ViewModels;
 
-namespace Restaurant.Client.UnitTests.ViewModels
+namespace Restaurant.Core.UnitTests.ViewModels
 {
 	[TestFixture]
 	public class SignInViewModelTest : BaseAutoMockedTest<SignInViewModel>
